refactor: 예외 계층 구조 수정 - 최하위 예외 계층 제거 (#146)
