+++ server/src/main/java/com/exchangediary/diary/service/DiaryCommandService.java
@@ -5,7 +5,7 @@ import com.exchangediary.diary.domain.entity.Diary;
 import com.exchangediary.diary.domain.entity.UploadImage;
 import com.exchangediary.diary.ui.dto.request.DiaryRequest;
 import com.exchangediary.global.exception.serviceexception.duplicate.DiaryDuplicateException;
-import com.exchangediary.global.exception.serviceexception.internalservererror.FailedImageUploadException;
+import com.exchangediary.global.exception.serviceexception.FailedImageUploadException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
