+++ server/src/main/java/com/exchangediary/global/exception/serviceexception/InvalidDateException.java
@@ -1,0 +1,9 @@
+package com.exchangediary.global.exception.serviceexception;
+
+import com.exchangediary.global.exception.ErrorCode;
+
+public class InvalidDateException extends ServiceException {
+    public InvalidDateException(String message, String value, ErrorCode errorCode) {
+        super(message, value, errorCode);
+    }
+}
