+++ server/src/main/java/com/exchangediary/global/exception/serviceexception/KakaoLoginFailureException.java
@@ -1,0 +1,9 @@
+package com.exchangediary.global.exception.serviceexception;
+
+import com.exchangediary.global.exception.ErrorCode;
+
+public class KakaoLoginFailureException extends ServiceException {
+    public KakaoLoginFailureException(String message, String value, ErrorCode errorCode) {
+        super(message, value, errorCode);
+    }
+}
