--- server/src/main/java/com/exchangediary/global/exception/serviceexception/duplicate/DiaryDuplicateException.java
@@ -1,9 +1,0 @@
-package com.exchangediary.global.exception.serviceexception.duplicate;
-
-import com.exchangediary.global.exception.ErrorCode;
-
-public class DiaryDuplicateException extends DuplicateException {
-    public DiaryDuplicateException(String value) {
-        super(value, ErrorCode.DIARY_DUPLICATED);
-    }
-}
