--- server/src/main/java/com/exchangediary/global/exception/serviceexception/internalservererror/KakaoTokenFailureException.java
@@ -1,9 +1,0 @@
-package com.exchangediary.global.exception.serviceexception.internalservererror;
-
-import com.exchangediary.global.exception.ErrorCode;
-
-public class KakaoTokenFailureException extends KakaoLoginFailureException {
-    public KakaoTokenFailureException(String value) {
-        super(value, ErrorCode.FAILED_TO_ISSUE_TOKEN);
-    }
-}
