--- server/src/main/java/com/exchangediary/global/exception/serviceexception/invliadrange/DateRangeException.java
@@ -1,9 +1,0 @@
-package com.exchangediary.global.exception.serviceexception.invliadrange;
-
-import com.exchangediary.global.exception.ErrorCode;
-
-public class DateRangeException extends InvalidRangeException {
-    public DateRangeException(String value) {
-        super(value, ErrorCode.INVALID_DATE_RANGE);
-    }
-}
