--- server/src/main/java/com/exchangediary/global/exception/serviceexception/notfound/UploadImageNotFoundException.java
@@ -1,9 +1,0 @@
-package com.exchangediary.global.exception.serviceexception.notfound;
-
-import com.exchangediary.global.exception.ErrorCode;
-
-public class UploadImageNotFoundException extends NotFoundException {
-    public UploadImageNotFoundException(String value) {
-        super(value, ErrorCode.UPLOAD_IMAGE_NOT_FOUND);
-    }
-}
